Fix filter in relation
Why:
Filter in relation broken when the relation name is not
the same with related tabel name, e.g.,
```
Post = Bookshelf.Model.extend({
  references() {
    return this.hasMany('Reference');
  }
});

Reference = Bookshelf.Model.extend({
  tableName: 'ref',
  post() {
    return this.belongsTo('Post')
  },
  author() {
    return this.belongsTo('Author');
  },
})

//
posts?filter[references.author.name][contains]=john&include=references
```

This change addresses the need by:
Changing the way to build joins in relation. Before it used the relation
name to match the parent, which is not accurate as the `parent` is filled
with `parent` table name, to the actual related table name.