Fix bug #9, multiple filters for a field
Why:
The plugin can not process multiple filters for a field.
For example
```
posts?filter[published][gte]=2017-05-01 00:00:00&filter[published][lte]=2017-05-03 08:00:00
```
the plugin will only process either the `gte` or `lte` depending on
which one of both is sorted first in the parsed query parameter
(usually ordered alphabetically).

This change addresses the need by:
Adding a process to catch such case --to _denormalize_ the object-- so
that the parser parses the query to two different filter:
```
[
  ['published', { gte: '2017-05-01 00:00:00' }],
  ['published', { lte: '2017-05-03 08:00:00' }],
]
```